AoC 2024 D21
This one was rough, and caused me to take a break. :-)

The main observation to optimize for part 2 is to note that, to press a button,
each directional pad must press A, and so all child pads will be positioned at A
at that moment (we need not save the whole button state); memoization is then
needed to take us past the finish line (though the sub-problems can be pretty
tricky to visualize in one's head).

This was a fun problem!

Also adds compile_flags.txt files for great benefit of clangd.